+++ src/query/executor/data_processing/graph_traversal/traverse.rs
@@ -415,7 +415,7 @@ impl<S: StorageEngine> ExecutorLifecycle for TraverseExecutor<S> {
 }
 
 impl<S: StorageEngine> ExecutorMetadata for TraverseExecutor<S> {
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.base.id
     }
