+++ src/query/executor/data_processing/join/right_join.rs
@@ -215,7 +215,7 @@ impl<S: StorageEngine + Send> ExecutorLifecycle for RightJoinExecutor<S> {
 }
 
 impl<S: StorageEngine + Send> ExecutorMetadata for RightJoinExecutor<S> {
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.base.id()
     }
