+++ src/query/executor/data_processing/set_operations/union.rs
@@ -26,7 +26,7 @@ pub struct UnionExecutor<S: StorageEngine> {
 impl<S: StorageEngine> UnionExecutor<S> {
     /// 创建新的Union执行器
     pub fn new(
-        id: usize,
+        id: i64,
         storage: Arc<Mutex<S>>,
         left_input_var: String,
         right_input_var: String,
@@ -110,7 +110,7 @@ impl<S: StorageEngine + Send + 'static> ExecutorLifecycle for UnionExecutor<S> {
 }
 
 impl<S: StorageEngine + Send + 'static> ExecutorMetadata for UnionExecutor<S> {
-    fn id(&self) -> usize {
+    fn id(&self) -> i64 {
         self.set_executor.id()
     }
